+++ CHANGELOG.md
@@ -10,11 +10,13 @@
 - `Innmind\Http\Sender::__invoke()` now returns `Innmind\Immutable\Attempt<Innmind\Immutable\SideEffect>`
 - `Innmind\Http\Factory\CookiesFactory` is now a final class
 - `Innmind\Http\Factory\EnvironmentFactory` is now a final class
+- `Innmind\Http\Factory\FilesFactory` is now a final class
 
 ### Removed
 
 - `Innmind\Http\Factory\Cookies\CookiesFactory`
 - `Innmind\Http\Factory\Environment\EnvironmentFactory`
+- `Innmind\Http\Factory\Files\FilesFactory`
 
 ### Fixed